ESRGAN/requirements.txt
appdirs==1.4.4
astunparse==1.6.3
cachetools==5.5.2
certifi==2019.11.28
chardet==3.0.4
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
etils==1.5.2
flatbuffers==25.2.10
fonttools==4.58.0
gast==0.4.0
gitdb==4.0.12
google-auth==2.40.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.58.0
grpcio==1.71.0
h5py==3.13.0
idna==2.8
importlib_metadata==8.7.0
importlib_resources==6.5.2
keras==2.11.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.9.4
numpy==1.26.4
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==25.0
pathtools==0.1.2
Pillow==9.5.0
promise==2.3
protobuf==3.19.6
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyYAML==6.0
requests==2.22.0
requests-oauthlib==2.0.0
requests-unixsocket==0.2.0
rsa==4.9.1
scipy==1.10.1
sentry-sdk==2.29.1
setproctitle==1.3.6
six==1.14.0
smmap==5.0.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-datasets==4.8.3
tensorflow-estimator==2.11.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.12.0
termcolor==3.1.0
toml==0.10.2
tqdm==4.65.0
typing_extensions==4.13.2
urllib3==2.4.0
wandb==0.15.4
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
